Migrate to using kubebuilder.
Note that the authorization server use case is a bit tricky as
kubebuilder assumes that the Kubernetes API Server is available.
To circumvent this, I kept a portion of the old webhook server
code to function as a standalone server.

Most of the new files are generated by kubebuilder, as is the
directory structure. This includes the Dockerfile and Makefile.

Also, secrets should be automatically generated by the server
and the webhooks automatically registered (in the non-standalone
case).

Apologies for the massive change but there wasn't a gradual way
to do this.

Signed-off-by: Max Smythe <smythe@google.com>