Initial copy of the mutation controller
Made a copy of the constraints controller, and changed all instances of
"Constraint" to "Mutation." This means things don't compile yet, so
commented out the implementation. Ensured that everything compiles.
Actual changes will come in the following commits; this is just to make
sure that future diffs will be easier to read.

Signed-off-by: Cameron McCarthy <camm@google.com>