Use patch to set finalizers
For pods in terminating state, their deletion timestamp is constantly
increasing, I'm not sure why but I suspect its do to with gatekeeper. As
a result, their resource version increases. This means that when we try
to remove the finalizer, there is an error 'this object has already
changed'. I think that using a patch will fix this.

Signed-off-by: Jack Kleeman <jackkleeman@gmail.com>