Add readiness tracker to ensure caches have been loaded before serving traffic
This commit adds readiness.Tracker, responsible for managing
expectations and observations of templates, constraints, and data.
Expectations are set according to the initial state of the cluster when
Gatekeeper starts - pre-existing templates and their corresponding
constraints are tracked until they are loaded by their respective
controllers into OPA. Similarly, cached data resources specified by
the Config singleton are also tracked to completion.

Templates or configured resource removed during warm-up are removed from
the initial expected set and will not block readiness.

The tracker is registered as a readiness probe - traffic will not be
routed to the admission controller while the readiness probe is not
returning success.

Fixes: #405
Signed-off-by: Oren Shomron <shomron@gmail.com>