Update README.md
In line number 502, to find the error, we should get the status of the template file, not the constraint. The template file is the one containing the Rego errors in the status while the constraint file was not accepted to the API yes. Running `kubectl apply -f` against the constraint file will result in the same error message `error: unable to recognize..."