Workaround for CRD parsing failure where envtest picks up kustomization.yaml
A kustomization.yaml was
[added](https://github.com/open-policy-agent/frameworks/commit/aa59bd1365711ea7a1978b880a0a2d337be154da)
in the constraints framework deployment directory. This is read by the
CRD loading code of controller-runtime's envtest and fails on parsing.
This leads to no CRDs being loaded for tests.

This commit works around the problem loading the framework CRDs
specifically where needed, rather than using the directory-based loading
code path.

Signed-off-by: Oren Shomron <shomron@gmail.com>