Decouple tracker from watch manager, use API server directly.
Resource are now listed directly from the API server in the readiness
tracker and will not be cached. This means the same resources will be
subsequently queried and cached as part of the normal controller
operation, but removes some coordination complexity.

Also add retry logic when tracking cached data, only short-circuit when
a resource kind is unregistered.

Signed-off-by: Oren Shomron <shomron@gmail.com>