Add readiness tracker to ensure caches have been loaded before serving traffic (#541)
* Add readiness tracker to ensure caches have been loaded before serving traffic

This commit adds readiness.Tracker, responsible for managing
expectations and observations of templates, constraints, and data.
Expectations are set according to the initial state of the cluster when
Gatekeeper starts - pre-existing templates and their corresponding
constraints are tracked until they are loaded by their respective
controllers into OPA. Similarly, cached data resources specified by
the Config singleton are also tracked to completion.

Templates or configured resource removed during warm-up are removed from
the initial expected set and will not block readiness.

The tracker is registered as a readiness probe - traffic will not be
routed to the admission controller while the readiness probe is not
returning success.

Fixes: #405
Signed-off-by: Oren Shomron <shomron@gmail.com>

* * Add circuit-breaker for readiness once expectations have been
  satisfied for the first time.
* Fix expected kinds according to ConstraintTemplate / Config rosters,
  only use those sets for checking satisfied expectations.
* Move config roster calculation into readiness package

Signed-off-by: Oren Shomron <shomron@gmail.com>

* Workaround for CRD parsing failure where envtest picks up kustomization.yaml

A kustomization.yaml was
[added](https://github.com/open-policy-agent/frameworks/commit/aa59bd1365711ea7a1978b880a0a2d337be154da)
in the constraints framework deployment directory. This is read by the
CRD loading code of controller-runtime's envtest and fails on parsing.
This leads to no CRDs being loaded for tests.

This commit works around the problem loading the framework CRDs
specifically where needed, rather than using the directory-based loading
code path.

Signed-off-by: Oren Shomron <shomron@gmail.com>

* Add test to verify config readiness with non-existent GVK references

Signed-off-by: Oren Shomron <shomron@gmail.com>

* Decouple tracker from watch manager, use API server directly.

Resource are now listed directly from the API server in the readiness
tracker and will not be cached. This means the same resources will be
subsequently queried and cached as part of the normal controller
operation, but removes some coordination complexity.

Also add retry logic when tracking cached data, only short-circuit when
a resource kind is unregistered.

Signed-off-by: Oren Shomron <shomron@gmail.com>

* Address remaining review comments

* Unexport Unsatisfied() method, remove from Expectations interface
* defer ExpectationsDone() where appropriate, avoids failures when
  listing resources cannot be completed (even after retries)
* Remove unneeded metav1.Object, runtime,Object method implementations
  on objKey
* Use RWMutex for synchronization in objectTracker
* Rename isRecoverable -> predicate for in retryLister

Signed-off-by: Oren Shomron <shomron@gmail.com>

* Create pkg/keys for shared keys used across packages. Fix comments.

Signed-off-by: Oren Shomron <shomron@gmail.com>

* Fix race condition registering readiness probe

Since controllers are registered asynchronously (waiting on certificate
generation), it was possible for the readiness probe registration call
to happen after the manager started, which caused an error and
CrashLoopBackoff. This commit moves probe registration back to the main
goroutine.

Signed-off-by: Oren Shomron <shomron@gmail.com>