Mutation: add a cache of mutations, sorted and thread safe.
The mutation system is a sorted cache of mutations, ordered by
- group
- kind
- namespace
- name

The mutation system is thread safe, and implement a rw lock to avoid blocking on multiple reads.
The number of writes is expected to be way less than the number of reads.

When applying the mutations, the system must converge after at most n+1 iterations, short
circuiting if converges earlier.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>