Cherry pick "Scope secrets cache to single namespace" (#972) to v3.2.x (#982)
* Allow multiple webhooks to be created (#882)

The PR prepares gatekeeper to be able to create multiple
webhooks. Common webhook code was extracted to a separate
file, and an update to cert rotator is used that allows
to update multiple webhooks.

Tested:
Gatekeeper with the changes was run. An incoming resource
was successfuly denied request due to gatekeeper policy.

Signed-off-by: mmirecki <mmirecki@redhat.com>
Signed-off-by: Oren Shomron <shomron@gmail.com>

Co-authored-by: Rita Zhang <rita.z.zhang@gmail.com>

* Scope secrets cache to single namespace (#972)

Incorporate open-policy-agent/cert-controller#16 to avoid caching
cluster-wide secrets in certificate rotation controller. This will
reduce the memory consumption of Gatekeeper on clusters with a large
number of secrets defined.

Fixes: #831

Signed-off-by: Oren Shomron <shomron@gmail.com>

Co-authored-by: Max Smythe <smythe@google.com>

* Fix e2e test flakiness (#964)

* Fix e2e test flakiness

Signed-off-by: Max Smythe <smythe@google.com>
Signed-off-by: Oren Shomron <shomron@gmail.com>

Co-authored-by: Marcin Mirecki <mmirecki@redhat.com>
Co-authored-by: Rita Zhang <rita.z.zhang@gmail.com>
Co-authored-by: Max Smythe <smythe@google.com>