Allow a tracker with observations but without expectations to be (#1062)
Satisfied

The object tracker maintains a list of all observed resources of it's
specified gvk.  These observations are removed whenever there is a
corresponding expectation for that resource.  The creation of an
expectation or an observation can come in any order for a resource.

It is possible that a resource can be observed, but that it is not
expected.  I.e. the resource is of the type tracked by the object
tracker, but is not one that should block the webhook.

In this case, we should not prevent the webhook from serving.  The
resource should be seen but not be cause to block the webhook.

Fixes #1023

Signed-off-by: juliankatz <juliankatz@google.com>