Upgrade controller-gen to 0.5.0
In support of the move to v1 CRDs, we will need up-to-date generation
tooling.  Gatekeeper was still on controller-gen 0.3.0, so this PR
updates that to 0.5.0.

One notable detail here is the removal of the caBundle field from the
webhook configurations.  This is not a mistake.  This field was related
to a bug in k8s that is now fixed, and thus the field was removed from
the generated output in controller-gen 0.4.1.

See kubernetes-sigs/controller-tools#495 for more info.

This PR contributes to open-policy-agent/gatekeeper#550

Signed-off-by: juliankatz <juliankatz@google.com>