Upgrade v1beta1 CRD yamls to v1
This PR changes the flags passed to controller-gen to generate v1 CRDs
for the CRDs defined in this repository.

It also includes changes from open-policy-agent/frameworks#114, which
update the ConstraintTemplate CRD to v1, and
open-policy-agent/frameworks#117, which allowed Gatekeeper to add the
Operations field to byPodStatus, which isn't built in to the byPodStatus
golang struct.

Contributes to #550

Signed-off-by: juliankatz <juliankatz@google.com>