Resolve rangeValCopy lint suggestions (#1292)
* Resolve rangeValCopy lint suggestions

While range loops are convenient ways of accessing each element in a
list, they can lead to immensely wasteful copying if not handled
carefully. In some cases, we're copying ~400 bytes just to access a
single field!

This PR fixes the issues found by the rangeValCopy linter. Usually, this
is resolved by indexing the list inline. In cases where the entire entry
is required, or many fields are accessed, I've created a variable which
creates a reference to the item in the list without copying it.

In one case, after the implicit shallow copy of the list, we then
explicitly create another shallow copy. I've made it so we only perform
a single shallow copy, as the code seems to intend.

Full list of fixes below. Each line includes how large the copied object
was. I've only done this change for objects larger than 128 bytes.

./pkg/audit/manager.go:273:3: rangeValCopy: each iteration copies 176 bytes (consider pointers or indexing)
./pkg/audit/manager.go:460:2: rangeValCopy: each iteration copies 176 bytes (consider pointers or indexing)
./pkg/audit/manager.go:560:2: rangeValCopy: each iteration copies 184 bytes (consider pointers or indexing)
./pkg/controller/constraintstatus/constraintstatus_controller.go:210:2: rangeValCopy: each iteration copies 376 bytes (consider pointers or indexing)
./pkg/controller/constrainttemplatestatus/constrainttemplatestatus_controller.go:181:2: rangeValCopy: each iteration copies 368 bytes (consider pointers or indexing)
./pkg/controller/mutatorstatus/mutatorstatus_controller.go:231:2: rangeValCopy: each iteration copies 376 bytes (consider pointers or indexing)
./pkg/readiness/ready_tracker.go:468:2: rangeValCopy: each iteration copies 384 bytes (consider pointers or indexing)
./pkg/readiness/ready_tracker.go:558:2: rangeValCopy: each iteration copies 353 bytes (consider pointers or indexing)
./pkg/upgrade/manager.go:120:2: rangeValCopy: each iteration copies 176 bytes (consider pointers or indexing)
./pkg/watch/manager_integration_test.go:452:3: rangeValCopy: each iteration copies 176 bytes (consider pointers or indexing)

Signed-off-by: Will Beason <willbeason@google.com>

* Remove shallow copy from ready_tracker.go

After studying the code, there appears to be no reason at all to perform
a shallow copy of the ConstraintTemplate. The list is not used for
anything else, so it is safe to use a reference to the item in the list.

Signed-off-by: Will Beason <willbeason@google.com>