Use "revive" instead of "golint" linter
Per the golint repository, the golint linter is no longer maintained.
The recommended golangci-lint replacement is revive, so this PR switches
us to use the new linter. This commit also includes all necessary
changes to make the new linter pass.

In configuring this I noticed the that our .golangci.yaml had a typo
which caused our misspell and lll configurations to be silently
discarded - I've corrected "linter" to "linters" and resolved the
spelling mistakes it found.

Also upgrade golangci-lint from 1.37.1 to 1.40.1 as previous versions
did not configure gosec properly.

Signed-off-by: Will Beason <willbeason@google.com>