Add structural information to Constraint Kind CRDs (#1249)
Kubernetes v1 CRDs require structural schemas.  This means that, where
possible, JSONSchemas should identify the type of a field.  Fields
without this type information will require the key: value pair of
x-kubernetes-unknown-fields: true to be set.  This signals the API
server to save the content found in that field to etcd without
validation.

This PR adds the remaining type information to the MatchSchema()
function, rendering its output structural.  This func's output is what
populates the `match` section of the CRD for a Constraint kind.

It also imports Constraint Framework, including the updates made in
open-policy-agent/frameworks#114.  This change transforms the schema
information provided by users in a v1beta1 ConstraintTemplate to be
structural.

Contributes to #550

Signed-off-by: juliankatz <juliankatz@google.com>