Fix manager image patching
The MANAGER_IMAGE_PATCH used to substitute a test image in the
gatekeeper.yaml was broken.  Only the first of the two patches was
applied, leaving one of the deployments un-patched.

This is likely related to a change in kustomize (my machine is running
v4.1.2).  Figuring out which version broke this change is less important
than fixing the problem, so I've gone ahead and done that.

Signed-off-by: juliankatz <juliankatz@google.com>