Upgrade constrainttemplate_controller to apiextensionsv1 (#1320)
This PR changes constrainttemplate_controller to use apiextensionsv1 to
create Constraint kind v1 CRDs.  That change revealed some additional
problems.  In particular, that byPod status was being pruned by the API
server.  This required a fix in Constraint Framework.

open-policy-agent/frameworks#120 adds a status field to the Constraint
kind CRD.  This allows Gatekeeper to add byPodStatus to the object.

This PR imports this new code and updates some tests.

Contributes to open-policy-agent/gatekeeper#550

Signed-off-by: juliankatz <juliankatz@google.com>