Lint code using dockerized golangci-lint
This means we don't need to manually install golangci-lint every time we
run our git workflow - we just use a generally-available one that
already has it installed.

It also means that "make lint" will work the same for everyone and the
workflow without additional coordination. Before this PR, it was
possible for maintainers to be on a different version of golangci-lint
than the one used by the GitHub workflow and so "make lint" would
provide inconsistent results.

Signed-off-by: Will Beason <willbeason@google.com>