Resolve PR comments
types.MakeID never needs to return an error. It can accept client.Object
instead of runtime.Object, which satisfies the desired interface.

Make parser.Path always a struct. The alternative was messy.

1) Performance-wise this is faster. Copying a parser.Path is faster than
creating and dereferencing pointers to it, especially as the type is
only 24 bytes.

2) Logic-wise, we always need a literal parser.Path to exist and don't
care to handle the case where parser.Path is nil since it wouldn't make
sense. This also means that in the case we use an empty Path, we'll
print a reasonable error rather than panicking.

Signed-off-by: Will Beason <willbeason@google.com>