Use MaxInt32 instead of MaxInt64
MaxInt64 breaks on 32-bit systems unless explicitly typecast to an
int64.

In this case, we do not expect users will ever need to reference nodes
more than 2 billion layers deep with gatekeeper. Thus, it feels safe to
make this assumption and use MaxInt32 as the starting value for the
minimum. We will encounter other problems long before this gives us
problems.

Signed-off-by: Will Beason <willbeason@google.com>