Parse gatekeeper-test flags (#1379)
This PR sets up several function and type stubs to be worked on in the
future. The intent here is to make it easy for development to be able to
happen in parallel PRs. The comments on these stubs is to clarify what
they _should_ do when complete rather than what they currently do.

These are the code paths which can be worked on in parallel:

1) Listing the set of files to test
2) Creating the Filter to select suites and tests to run
3) Running test suites and printing test output

Fixes: #1369

Signed-off-by: Will Beason <willbeason@google.com>