Add PreserveUnknownFields: false to CRDs (#1356)
* Add PreserveUnknownFields: false to CRDs

First commit, adds PreserveUnknownFields: false to the non-mutation
CRDs.  The kustomization pipeline is in a bit of a weird state given it
being in alpha, so I'm not sure what else to add for that yet.

Signed-off-by: juliankatz <juliankatz@google.com>

* Preserve --> preserve

Signed-off-by: juliankatz <juliankatz@google.com>

* Fix mutation CRDs kustomization

The patching was in the wrong place.  Per Rita's comment, I moved it
from:

config/overlays/mutation/kustomization.yaml

to

config/overlays/mutation_webhook/kustomization.yaml

Now the CRDs are actually being kustomized.

Signed-off-by: juliankatz <juliankatz@google.com>

* Add PreserveUnknownFields: false to CRDs

First commit, adds PreserveUnknownFields: false to the non-mutation
CRDs.  The kustomization pipeline is in a bit of a weird state given it
being in alpha, so I'm not sure what else to add for that yet.

Signed-off-by: juliankatz <juliankatz@google.com>

* Preserve --> preserve

Signed-off-by: juliankatz <juliankatz@google.com>

* Add load_restrictor LoadRestrictionsNone in deploy-mutation

Signed-off-by: juliankatz <juliankatz@google.com>

Co-authored-by: Sertaç Özercan <852750+sozercan@users.noreply.github.com>