parse Templates + Constraints in Suites
Make "gator test" parse ConstraintTemplates and Constraints specified by
Suites. The focus is on:
1) Creating the OPA client
2) Compiling Templates
3) Instantiating Constraints

When errors occur here we don't return an error - we return an
ErrorResult which will eventually be printed in a way similar to test
failures and not as a runtime error executing tests.

There's several error paths we can't easily test - I've marked those
with comments.

This PR does not actually run test Suites - that functionality would
easily double the length of this already-large PR.

Note that we do not simply exist immediately if a Suite contains no
Tests - we support Suites which define no tests, in which case we just
compile the Template and Constraint, returning a failure if either fails
to be added to the OPA client.

We should be cautious when adding tests to suite_test.go. Each takes
about 100ms due to the significant costs of creating an OPA client,
compiling a Template, and then instantiating a Constraint. There is no
good way around this as part of the behavior we need to test is the
ability to trigger and catch errors from a real OPA client.

Signed-off-by: Will Beason <willbeason@google.com>