Add integer keyValue support to mutation path parser / mutators (#1394)
* Add integer keyValue support to mutation path parser / mutators.

Previously, the mutation path parser only supported string types for
keyValues selecting items in a list. This PR adds support for integers,
such that we can now express paths like this:

```
spec.containers[name: opa].ports[containerPort: 8888].name
```

Integers are assumed to be represented in decimal and are interpreted as
positive int64 values.

The mutation code was updated accordingly to match based on type.
Additionally, string serialization of paths was updated to be more
selective on which strings are quoted.

Fixes #1355

* Add additional tests, resolve lint warnings.

* Additional test cases for digits in unquoted identifiers

* Fixed additional parsing  roundtripping cases

Signed-off-by: Oren Shomron <shomron@gmail.com>
Signed-off-by: juliankatz <juliankatz@google.com>