Read test files (#1412)
Implement functionality for reading test Suites. This includes reading
single file, reading suites in a single directory, and reading suites in
a directory and its subdirectories.

Note that it is intentional that the Suite type is still empty - a
parallel CL will begin adding fields and functionality to it. This CL is
just focused on reading Suites from the disk.

I've opted to not test certain edge cases - for example permission
errors - as we'd either have to define our own fs.FS implementation or
have unit tests that write/read from disk. As-is, unit tests cover the
core logic paths and 90% of the code.

Signed-off-by: Will Beason <willbeason@google.com>