Print output of tests similar go go test
A first pass at printing output similar to what is shown for "go test".

Sadly we can't just use testing.T without asking users to write their
own unit tests. There is no accessible printer that automatically does
this formatting for us, so we must replicate it.

This does not print output as tests are run even though the result
output appears to suggest this. In the future we will do so, but as we
don't yet run tests that would be premature. The point here is to make
the output consistent with "go test".

Signed-off-by: Will Beason <willbeason@google.com>