Use future-compatible result types
As stated elsewhere, the purpose of this PR is NOT the result types.
They are not final and merely exist as stubs. They will be given a
complete treatment in a dedicated PR.

Similarly, this PR does not handle printing output. There will be a PR
for this, and the as-is state of not printing output is not final.

Signed-off-by: Will Beason <willbeason@google.com>