Benchmark AssignMutator.Mutate code
Add benchmark tests for AssignMutator.Mutate.

In writing this, I found that mutating objects takes about 10x longer
than not mutating objects. Profiling reveals that we spend over 90% of
compute time in AssingMutator.Value(), where we convert from the raw
JSON value to a literal Go struct.

In a separate PR I'll suggest a change that resolves this performance
issue. Most likely we'll want to cache the assigned value. We can then
either reflect.DeepCopy the value or assign the cached value directly,
depending on whether we feel it safe.

Signed-off-by: Will Beason <willbeason@google.com>