Refactor run code into Runner
Now that much of the test running logic is defined, it is more apparent
how to organize it.

Code specific to running tests, but which does not change with the
nature of the tests being run or how the results are printed, should
live in its own place. It may be decorated in the future with objects
(like Printers) which are specified by the user.

This PR has zero changes to logic - it only moves logic around.

Signed-off-by: Will Beason <willbeason@google.com>