deepcopy cached value instead of unmarshalling json (#1439)
Instead of keeping the assigned value for AssignMutator as json, cache the unmarshalled value.

Unmarshalling JSON is incredibly expensive, even for simple values. We should avoid this when possible. Instead, we should cache the value to assign, and then deepcopy it when needed. Fortunately, the runtime package provides us a method just for this!

This change results in a ~90% reduction in compute time for most common uses. The existing benchmarks are for assigning a string, but the speedup is even more for complex assigned values.

Signed-off-by: Will Beason <willbeason@google.com>