[gk-test] Run case (#1446)
* Print output of tests similar go go test

A first pass at printing output similar to what is shown for "go test".

Sadly we can't just use testing.T without asking users to write their
own unit tests. There is no accessible printer that automatically does
this formatting for us, so we must replicate it.

This does not print output as tests are run even though the result
output appears to suggest this. In the future we will do so, but as we
don't yet run tests that would be premature. The point here is to make
the output consistent with "go test".

Signed-off-by: Will Beason <willbeason@google.com>

* Run test cases

Add logic for running individual test cases. Now that we've compiled the
constraints, all we need to do is Review() the specified objects and
compare the Results with what the user specified.

This PR does not add Assertion functionality - we'll do that for alpha,
not pre-alpha.

Also add unit tests that verify we are, in fact, running the test cases.

Signed-off-by: Will Beason <willbeason@google.com>