Add frameworks apis to scheme
I recently ran `make vendor` on gatekeeper master and noticed that a
number of files from frameworks/constraint were being deleted.  Why?

It runs out that a recent PR (#1445) removed a dependency on
frameworks/constraint/pkg/apis and `make vendor` was pruning the unused
code from the vendor/ directory.

As the API types declared in frameworks/constraint are part of the API
used in g8r, I thought it appropriate to add them to the scheme.  This
fixes the problem with make vendor.

Signed-off-by: juliankatz <juliankatz@google.com>