Add frameworks apis to scheme
I recently ran `make vendor` on gatekeeper master and noticed that a
number of files from frameworks/constraint were being deleted.  Why?

It runs out that a recent PR (#1445) removed a dependency on
frameworks/constraint/pkg/apis and `make vendor` was pruning the unused
code from the vendor/ directory.

Further, while we might expect the AddToScheme calls to be used when we
setup the scheme in main.go, it turns out that we were making
essentially duplicate versions of the AddToScheme files in g8r.  I've
replaced these with a single file that calls all the AddToScheme
functions in frameworks/constraint.

Signed-off-by: juliankatz <juliankatz@google.com>