more TODOs fixed
Part of the difficulty in this work came from defer functions in tests.
These aren't properly scoped, so when they panic or fail, the failure is
global and not attributed to the test.

I'll make a PR in the future which converts our use of defer in tests to
t.Cleanup.

Signed-off-by: Will Beason <willbeason@google.com>