Make Context usage consistent (#1457)
* Make Context usage consistent

In trying to upgrade our controller-runtime dependency to v0.9.0, I
found that the upgrade introduced several bugs. controller-runtime is
now being more thorough in its usage of Context, which meant we had
cases where we _had_ a context, but weren't properly passing it down the
call chain.

This CL eliminates all usage of context.Background() from pkg/, which
does resolve that part of the upgrade issue.

The other issue (to be resolved separately) is that we both depend on
the controller-runtime package in vendor/ and a different version in
third_party/.

Signed-off-by: Will Beason <willbeason@google.com>

* remove usage of context.TODO()

Signed-off-by: Will Beason <willbeason@google.com>

* more TODOs fixed

Part of the difficulty in this work came from defer functions in tests.
These aren't properly scoped, so when they panic or fail, the failure is
global and not attributed to the test.

I'll make a PR in the future which converts our use of defer in tests to
t.Cleanup.

Signed-off-by: Will Beason <willbeason@google.com>

* Don't use real contexts for metrics

Opencensus asks for context, but doesn't really want to use it like a
real cancellable context. So instead we make contexts as needed to pass
as arguments.

Signed-off-by: Will Beason <willbeason@google.com>

* Remove many unused contexts

While doing this I had to turn on unparam, which led me to removing
several other unused parameters to reduce noise while looking for unused
contexts.

Signed-off-by: Will Beason <willbeason@google.com>

* Per reviewer comment test no parent ctx

For replay.go

Signed-off-by: Will Beason <willbeason@google.com>