Refactor mutator Matches() to make extension easy (#1494)
Refactor mutator Matches() to make extension easy

As part of #1402, I'm tasked with adding Name to the match criteria for
mutators.  Unfortunately, I found this to be prohibitively difficult to
implement with the current logic.

Fortunately, the logic was well tested and defined in the docs, which
say:

"Note that if multiple matchers are specified, a resource must satisfy
each top-level matcher (kinds, namespaces, etc.) to be in scope"

from: https://open-policy-agent.github.io/gatekeeper/website/docs/howto

Knowing that, I knew I could refactor this logic to AND a series of
matching functions.  So, I've separated that logic into distinct
sections that are ANDed together.  This makes adding a new field (which
I'll do in a subsequent PR) very easy.

Contributes to #1402

Signed-off-by: juliankatz <juliankatz@google.com>