Implement violations
Users may now specify:

- that an object is expected to produce violations
- the expected number of violations
- expected violation messages

This deviates from the design by making Violations an object rather than
a list - making it an array was awkward to work with. I've tried to
leave room so that it is reasonably extensible. I think we'll have a
better idea of the ideal structure when we actually get more assertions
we want to allow users to make.

Signed-off-by: Will Beason <willbeason@google.com>