Run gator test 2 (#1519)
* start refactoring per suggestions

Signed-off-by: Will Beason <willbeason@google.com>

* fix object parameter usage

Signed-off-by: Will Beason <willbeason@google.com>

* Implement violations

Users may now specify:

- that an object is expected to produce violations
- the expected number of violations
- expected violation messages

This deviates from the design by making Violations an object rather than
a list - making it an array was awkward to work with. I've tried to
leave room so that it is reasonably extensible. I think we'll have a
better idea of the ideal structure when we actually get more assertions
we want to allow users to make.

Signed-off-by: Will Beason <willbeason@google.com>

* Add example violations to ReadSuite tests

Note that users can specify simply that there are violations by setting
"violations: {}" and leaving it as an empty object, and they don't have
to specify anything further.

Signed-off-by: Will Beason <willbeason@google.com>

* Run golangci-lint run --fix

Signed-off-by: Will Beason <willbeason@google.com>

* Responds to nits

Also refactor runner to reduce duplicate code and reduce responsibility
of individual functions.

Signed-off-by: Will Beason <willbeason@google.com>

* Implement reviewer-suggested changes

This more closely follows the original design, with the explicit changes
discussed in the gatekeeper weekly meeting.

Signed-off-by: Will Beason <willbeason@google.com>

* Add comprehensive unit tests for asssertions

This covers a reasonable variety of use cases for Assertions.

Signed-off-by: Will Beason <willbeason@google.com>

* Clarify Assertions field default behavior.

Signed-off-by: Will Beason <willbeason@google.com>

* Use k8s IntOrStr

Since IntOrStr doesn't actually support YAML, we have to make a pass
through JSON in order for this to work.

Signed-off-by: Will Beason <willbeason@google.com>

* Add test case for when string value is invalid

Signed-off-by: Will Beason <willbeason@google.com>

* Improve Assertions field comments

Signed-off-by: Will Beason <willbeason@google.com>

* Resolve review comments

- Rename tests
- Refactor error messages to reduce variety of test failure messages

Signed-off-by: Will Beason <willbeason@google.com>

* Fix lint errors

Signed-off-by: Will Beason <willbeason@google.com>