Add name match support
In #1402, a user sought to modify an object of a specific name.
Previously, there was no way to do this.  The user attempted to use
assignIf against metadata.name, but assignIf only allows a mutation to
check against the current value of the field to be mutated.

This PR adds exact name matching support, as well as prefix-based glob
matching.

Contributes to #1402

Signed-off-by: juliankatz <juliankatz@google.com>