Ensure helm hooks run as non-root and make PSP optional
Helm hooks now run as non-root user which will make sure the hooks
run on clusters which constrain Pods (e.g. using PodSecurityPolicies).
Also make PodSecurityPolicy optional (a) because PSP are deprecated
in Kuberntes 1.21 and (b) because it is a bad practice to (unconditionally)
create PSP in helm charts (or kustomizations) as PSP are usually managed
by cluster operators and not application operators.

Signed-off-by: Alexander Berger <alex-berger@gmx.ch>