Preserve conflicting mutators (#1569)
Preserve conflicting mutator paths, making behavior independent of the order in which conflicting mutators are added.

Increase test coverage from 70% to 89%. Added tests are unit tests that fake out the dependencies of Reconciler, so they don't require the same setup as the existing integration tests in the same package.

* Start refactoring system code

Signed-off-by: Will Beason <willbeason@google.com>

* Add tests for untested paths

Signed-off-by: Will Beason <willbeason@google.com>

* Simplify Mutate

Signed-off-by: Will Beason <willbeason@google.com>

* Keep track of mutation ids for ordering

We don't need two copies of the same information, and it makes testing
easier this way.

Signed-off-by: Will Beason <willbeason@google.com>

* Rename to orderedIDs for clarity

Since it isn't holding mutators any more.

Signed-off-by: Will Beason <willbeason@google.com>

* Add structured failed mutation error

Signed-off-by: Will Beason <willbeason@google.com>

* modify reconciler and start updating tests

Signed-off-by: Will Beason <willbeason@google.com>

* Finish testing reconciler base paths

Signed-off-by: Will Beason <willbeason@google.com>

* Continue testing reconciler edge cases

Signed-off-by: Will Beason <willbeason@google.com>

* Change Pod Ownership to member variable

As a global variable, with multiple tests manipulating the value, it
makes unit tests dependenton each other.

Signed-off-by: Will Beason <willbeason@google.com>

* Finish testing Reconciler edge cases

We properly preserve errors now

Signed-off-by: Will Beason <willbeason@google.com>

* Fix test

The problem was my test code. Now the reconciler edge cases work.

Signed-off-by: Will Beason <willbeason@google.com>

* Add test for mutation webhook

The logic already works, so no change to non-test code required

Signed-off-by: Will Beason <willbeason@google.com>

* Make GVK a field of reconciler

This isn't expected to change dynamically through the runtime of the
program, so just set it at Reconciler creation

Signed-off-by: Will Beason <willbeason@google.com>

* Fix incorrect comment on method

Signed-off-by: Will Beason <willbeason@google.com>

* Run make manifests

Signed-off-by: Will Beason <willbeason@google.com>

* Ensure we fix mutator errors when HasDiff is false

Signed-off-by: Will Beason <willbeason@google.com>

* Handle failure to add conflict error to PodStatus

This ensures we properly update PodStatus in the case we successfully
modify the Schema to include the conflict, but then fail to update the
PodStatus. In the second Reconcile of the conflicting Mutator, we must
return the schema conflict error rather than nil to indicate that the
PodStatus still needs to be updated.

Signed-off-by: Will Beason <willbeason@google.com>

* golangci-lint run --fix

Signed-off-by: Will Beason <willbeason@google.com>

* Better explain methods with comments

Deduplicate logic for reporting conflicts

Signed-off-by: Will Beason <willbeason@google.com>

* Explain behavior of conditional error updating

Signed-off-by: Will Beason <willbeason@google.com>

* Add controller tests

Fix controller tests which were logging noisy errors by specifying
metadata.namespace for the Pod.

Signed-off-by: Will Beason <willbeason@google.com>

* golangci-lint run --fix

Signed-off-by: Will Beason <willbeason@google.com>

* Fix nits

Signed-off-by: Will Beason <willbeason@google.com>

* Add more edge cases

Cover a few more edge cases in reconciler unit tests

Simplify redundant reconciler logic.

Signed-off-by: Will Beason <willbeason@google.com>

* golangci-lint run --fix

Signed-off-by: Will Beason <willbeason@google.com>

* Fix PR comments

Signed-off-by: Will Beason <willbeason@google.com>