Remove PodOwnership option (#1585)
* Remove PodOwnership option

This was introduced for tests, but is unnecessary. All that is needed
for setting ownerreferences to work properly is to set the UID of fake
Pods used in testing.

This is because ownerreference validation requires that UID be set on
the reference (to deduplicate objects with identical names/namespaces
which existed at different points in time).

This raises the point that we should probably have a general-purpose
method for instantiating a fake Pod for use in testing and debugging. It
isn't obvious to me where that is. Feel free to suggest a location. Do
note that it would be in production code since the debug setting is
compiled in non-test files.

Signed-off-by: Will Beason <willbeason@google.com>

* Fix Cache test

Signed-off-by: Will Beason <willbeason@google.com>

* Make Cache test less flaky

Signed-off-by: Will Beason <willbeason@google.com>

* Use fakes for default-instantiating objects

Signed-off-by: Will Beason <willbeason@google.com>

* Deepcopy instance inside of ensureX commands

Also split ensureDeleted out of ensureCreated.

Signed-off-by: Will Beason <willbeason@google.com>

* Resolve reviewer nits

Signed-off-by: Will Beason <willbeason@google.com>