Make gator compatible with Windows
In general, remove all usage of "filepath" except when converting to
slash-delimited filepaths. The fs.FS interface always requires
Unix-style paths, so we need to always convert to slash-delimited paths.

The problem we run into is that a relative Windows path of multiple elements
is a valid filename for a file on a Unix filesystem. Thus, if we allow either
then we make whether suites work dependent on the machine executing them. This
doesn't make sense since many dev teams are cross-platform. Instead, we'll
assume any paths in suite files are slash-delimited.

Thus gator test will not work in the case of suites containing Windows-style
paths. This doesn't feel like a dealbreaker, so I'll note it in documentation.
Also note that it means paths containing escaped backslashes will not be
interpreted correctly on Windows machines. While that is platform-specific
behavior we can deal with that in a separate PR.

I've added a conversion to slash-delimited for the path passed directly
to the CLI to make it easy for users who are copy+pasting paths on Windows.
This conversion only occurs when running on Windows so non-Windows users are
not impacted.

I've verified these changes make "gator test" work on a Windows machine.

Never assume compatibility: always test on a real Windows machine.

Signed-off-by: Will Beason <willbeason@gmail.com>